+++ src/frameworks/core_animation/ca_eagl_layer.rs
@@ -91,6 +91,9 @@ pub fn find_fullscreen_eagl_layer(env: &mut Environment) -> id {
                 })
             || layer_host_obj.hidden
             || layer_host_obj.opacity != 1.0
+            // TODO: support affine transforms that result in a full-screen
+            //       layer (typical example is 90° rotation).
+            || !layer_host_obj.affine_transform.is_identity()
         {
             return nil;
         }