+++ tests/TestApp_source/GUITestsCALayerTestsView.m
@@ -5,25 +5,29 @@
  */
 
 #include "system_headers.h"
+#include <math.h>
 
 #include "GUITestsCALayerTestsView.h"
 
-#define NUM_TESTS 3
+#define NUM_TESTS 12
 
 @implementation GUITestsCALayerTestsView : UIView
 
-UILabel *label;
+UILabel *title;
 UIView *testArea;
 NSUInteger testNum;
+UIView *lastTappedView;
+UILabel *lastTappedLocalFrameLabel;
+UILabel *lastTappedGlobalFrameLabel;
 
 - (instancetype)initWithFrame:(CGRect)frame {
   [super initWithFrame:frame];
 
-  label = [[UILabel alloc] initWithFrame:[self bounds]];
-  label.text = [NSString stringWithUTF8String:"CALayer tests (press →)"];
-  label.textAlignment = UITextAlignmentCenter;
-  label.frame = CGRectMake(0, 0, 320, 20);
-  [self addSubview:label];
+  title = [[UILabel alloc] initWithFrame:[self bounds]];
+  title.text = [NSString stringWithUTF8String:"CALayer tests (press →)"];
+  title.textAlignment = UITextAlignmentCenter;
+  title.frame = CGRectMake(0, 0, 320, 20);
+  [self addSubview:title];
 
   UIButton *button1 = [UIButton buttonWithType:UIButtonTypeRoundedRect];
   [button1 setTitle:[NSString stringWithUTF8String:"←"]
@@ -45,6 +49,17 @@ NSUInteger testNum;
   [self addSubview:button2];
   [button2 layoutSubviews]; // FIXME: workaround for touchHLE not calling this
 
+  lastTappedLocalFrameLabel =
+      [[UILabel alloc] initWithFrame:CGRectMake(10, 340, 300, 20)];
+  lastTappedLocalFrameLabel.textColor = [UIColor whiteColor];
+  lastTappedLocalFrameLabel.backgroundColor = [UIColor clearColor];
+  [self addSubview:lastTappedLocalFrameLabel];
+  lastTappedGlobalFrameLabel =
+      [[UILabel alloc] initWithFrame:CGRectMake(10, 360, 300, 20)];
+  lastTappedGlobalFrameLabel.textColor = [UIColor whiteColor];
+  lastTappedGlobalFrameLabel.backgroundColor = [UIColor clearColor];
+  [self addSubview:lastTappedGlobalFrameLabel];
+
   // Don't display any test initially. The testing for convertPoint:toLayer: etc
   // won't produce the right results until this view has actually been added to
   // the window.
@@ -54,7 +69,9 @@ NSUInteger testNum;
 }
 
 - (void)dealloc {
-  [label release];
+  [title release];
+  [lastTappedLocalFrameLabel release];
+  [lastTappedGlobalFrameLabel release];
   [testArea release];
   [super dealloc];
 }
@@ -70,11 +87,14 @@ NSUInteger testNum;
   [self displayTest];
 }
 - (void)displayTest {
-  label.text = [NSString
+  title.text = [NSString
       stringWithFormat:[NSString stringWithUTF8String:"CALayer test %u/%u"],
                        testNum, NUM_TESTS];
   [testArea removeFromSuperview];
   [testArea release];
+  lastTappedView = nil;
+  lastTappedLocalFrameLabel.text = [NSString stringWithUTF8String:""];
+  lastTappedGlobalFrameLabel.text = [NSString stringWithUTF8String:""];
   testArea = [[UIView alloc] initWithFrame:CGRectMake(10, 30, 300, 300)];
   testArea.backgroundColor = [UIColor grayColor];
   [self addSubview:testArea];
@@ -85,10 +105,29 @@ NSUInteger testNum;
                                              testNum])];
 }
 
+- (void)tick {
+  if (lastTappedView) {
+    lastTappedLocalFrameLabel.text = NSStringFromCGRect(lastTappedView.frame);
+    lastTappedGlobalFrameLabel.text =
+        NSStringFromCGRect([testArea convertRect:lastTappedView.frame
+                                        fromView:lastTappedView]);
+  }
+  SEL tickSelector = NSSelectorFromString([NSString
+      stringWithFormat:[NSString stringWithUTF8String:"test%uTick"], testNum]);
+  if ([self respondsToSelector:tickSelector]) {
+    [self performSelector:tickSelector];
+  }
+}
+
 - (UIView *)addViewWithFrame:(CGRect)frame color:(UIColor *)color {
+  return [self addViewWithFrame:frame color:color superview:testArea];
+}
+- (UIView *)addViewWithFrame:(CGRect)frame
+                       color:(UIColor *)color
+                   superview:(UIView *)parent {
   UIView *view = [[UIView alloc] initWithFrame:frame];
   view.backgroundColor = color;
-  [testArea addSubview:view];
+  [parent addSubview:view];
   [view release];
   return view;
 }
@@ -102,7 +141,15 @@ NSUInteger testNum;
   return label;
 }
 
-// These tests should all look like three squares arranged diagonally.
+- (void)touchesBegan:(NSSet<UITouch *> *)touches withEvent:(UIEvent *)event {
+  CGPoint locationInView = [[touches anyObject] locationInView:testArea];
+  UIView *hitView = [testArea hitTest:locationInView withEvent:event];
+  hitView.alpha = 1.0 - (hitView.alpha - 0.5);
+  lastTappedView = hitView;
+}
+
+// These tests should all look like three squares arranged diagonally (in some
+// cases with animation).
 // The color differences make it more obvious when you've switched tests.
 
 - (void)test1 {
@@ -156,5 +203,133 @@ NSUInteger testNum;
   view3.layer.position = CGPointMake(200.0, 200.0);
   view3.layer.affineTransform = CGAffineTransformMakeTranslation(100.0, 100.0);
 }
+- (void)test4 {
+  [self test2];
+}
+- (void)test4Tick {
+  NSTimeInterval t = [[NSProcessInfo processInfo] systemUptime];
+  t -= floor(t);
+  for (NSUInteger i = 0; i < testArea.subviews.count; i++) {
+    [testArea.subviews objectAtIndex:i].transform =
+        CGAffineTransformMakeRotation(t * M_PI * 2);
+  }
+}
+- (void)test5 {
+  [self test2];
+}
+- (void)test5Tick {
+  NSTimeInterval t = [[NSProcessInfo processInfo] systemUptime];
+  t -= floor(t);
+  for (NSUInteger i = 0; i < testArea.subviews.count; i++) {
+    ((UIView *)[testArea.subviews objectAtIndex:i]).transform =
+        CGAffineTransformMakeScale(0.5 + (sin(t * M_PI * 2) / 4 + 0.25),
+                                   0.5 + (cos(t * M_PI * 2) / 4 + 0.25));
+  }
+}
+- (void)test6 {
+  UIView *view1 = [self addViewWithFrame:CGRectMake(0, 0, 100, 100)
+                                   color:[UIColor redColor]];
+  UIView *view2 = [self addViewWithFrame:CGRectMake(0, 0, 100, 100)
+                                   color:[UIColor greenColor]
+                               superview:view1];
+  view2.transform = CGAffineTransformMakeTranslation(100.0, 100.0);
+  UIView *view3 = [self addViewWithFrame:CGRectMake(0, 0, 100, 100)
+                                   color:[UIColor blueColor]
+                               superview:view2];
+  view3.transform = CGAffineTransformMakeTranslation(100.0, 100.0);
+}
 
+- (void)test7 {
+  UIView *view1 = [self addViewWithFrame:CGRectMake(0, 0, 100, 100)
+                                   color:[UIColor cyanColor]];
+  view1.bounds = CGRectMake(100.0, 100.0, 100.0, 100.0);
+  UIView *view2 = [self addViewWithFrame:CGRectMake(200, 200, 100, 100)
+                                   color:[UIColor magentaColor]
+                               superview:view1];
+  UIView *view3 = [self addViewWithFrame:CGRectMake(100, 100, 100, 100)
+                                   color:[UIColor yellowColor]
+                               superview:view2];
+}
+
+// These tests also use three squares, but they break the earlier rule that the
+// squares should always have the same arrangement; the transforms will shift
+// their positions.
+
+- (void)test8 {
+  UIView *view1 = [self addViewWithFrame:CGRectMake(0, 0, 100, 100)
+                                   color:[UIColor orangeColor]];
+  UIView *view2 = [self addViewWithFrame:CGRectMake(0, 0, 100, 100)
+                                   color:[UIColor greenColor]
+                               superview:view1];
+  view2.transform = CGAffineTransformRotate(
+      CGAffineTransformMakeTranslation(100.0, 100.0), M_PI / 8);
+  UIView *view3 = [self addViewWithFrame:CGRectMake(0, 0, 100, 100)
+                                   color:[UIColor purpleColor]
+                               superview:view2];
+  view3.transform = CGAffineTransformRotate(
+      CGAffineTransformMakeTranslation(100.0, 100.0), M_PI / 8);
+}
+- (void)test9 {
+  UIView *view1 = [self addViewWithFrame:CGRectMake(0, 0, 100, 100)
+                                   color:[UIColor cyanColor]];
+  UIView *view2 = [self addViewWithFrame:CGRectMake(0, 0, 100, 100)
+                                   color:[UIColor magentaColor]
+                               superview:view1];
+  view2.transform = CGAffineTransformScale(
+      CGAffineTransformMakeTranslation(100.0, 100.0), 0.75, 0.75);
+  UIView *view3 = [self addViewWithFrame:CGRectMake(0, 0, 100, 100)
+                                   color:[UIColor yellowColor]
+                               superview:view2];
+  view3.transform = CGAffineTransformScale(
+      CGAffineTransformMakeTranslation(100.0, 100.0), 0.75, 0.75);
+}
+
+// These tests don't have the three squares pattern.
+- (void)test10 {
+  [self addViewWithFrame:CGRectMake(100, 100, 100, 100)
+                   color:[UIColor blackColor]];
+  [self addViewWithFrame:CGRectMake(100, 100, 100, 100)
+                   color:[UIColor redColor]];
+}
+- (void)test10Tick {
+  NSTimeInterval t = [[NSProcessInfo processInfo] systemUptime];
+  t -= floor(t);
+  [testArea.subviews objectAtIndex:1].transform =
+      CGAffineTransformMakeRotation(t * M_PI * 2);
+  [testArea.subviews objectAtIndex:0].frame =
+      [testArea.subviews objectAtIndex:1].frame;
+}
+- (void)test11 {
+  [self addViewWithFrame:CGRectMake(100, 100, 100, 100)
+                   color:[UIColor redColor]];
+  UIView *view2 = [self addViewWithFrame:CGRectMake(100, 100, 100, 100)
+                                   color:[UIColor blackColor]];
+  view2.transform = CGAffineTransformMakeRotation(M_PI / 4);
+}
+- (void)test11Tick {
+  NSTimeInterval t = [[NSProcessInfo processInfo] systemUptime];
+  t -= floor(t);
+  [testArea.subviews objectAtIndex:0].transform = CGAffineTransformMakeScale(
+      1.25 + sin(t * M_PI * 2) / 4, 1.25 + sin(t * M_PI * 2) / 4);
+  [testArea.subviews objectAtIndex:1].frame =
+      [testArea.subviews objectAtIndex:0].frame;
+}
+- (void)test12 {
+  [self addViewWithFrame:CGRectMake(100, 100, 100, 100)
+                   color:[UIColor redColor]];
+  UIView *view2 = [self addViewWithFrame:CGRectMake(50, 50, 200, 200)
+                                   color:[UIColor clearColor]];
+  [self addViewWithFrame:CGRectMake(0, 0, 50, 50)
+                   color:[UIColor blackColor]
+               superview:view2];
+  [self addViewWithFrame:CGRectMake(150, 150, 50, 50)
+                   color:[UIColor blackColor]
+               superview:view2];
+}
+- (void)test12Tick {
+  NSTimeInterval t = [[NSProcessInfo processInfo] systemUptime];
+  [testArea.subviews objectAtIndex:1].transform = CGAffineTransformRotate(
+      CGAffineTransformMakeScale(0.5, 0.5), (M_PI / 2) * (((int)floor(t)) % 4));
+  [self test11Tick];
+}
 @end